requirements.txt
# use "pip show x" to see versions

streamlit==1.41.1
qdrant_client==1.13.2
llama_index==0.12.12
llama_index.llms.ollama==0.5.0
llama_index.embeddings.huggingface==0.5.1
llama_index.vector_stores.qdrant==0.4.3
docx2txt==0.8
torch==2.5.1
transformers==4.48.1
python-dotenv==1.0.1